requirements.txt
openai
sentence-transformers
numpy
python-dotenv
httpx==0.27.2